+++ src/Lib/Parser/AnimeParse.php
@@ -17,6 +17,9 @@ class AnimeParse extends TemplateParse
 
         $this->addRule('link_canonical', '~<link rel="canonical" href="(.*)" />~', function() {
            $this->model->set('Anime', 'link_canonical', $this->matches[1]);
+
+            preg_match('~myanimelist.net/(.+)/(.*)/~', $this->model->get('Anime', 'link_canonical'), $this->matches);
+            $this->model->set('Anime', 'mal_id', (int) $this->matches[2]);
         });
 
         $this->addRule('title', '~<h1 class="h1"><span itemprop="name">(.*)</span></h1>~', function() {
@@ -205,7 +208,7 @@ class AnimeParse extends TemplateParse
                         foreach ($links as $key2 => $value2) {
                             if (preg_match('~<a href="(.*)/(.*)/(.*)">(.*)(</a>|)~', $value2, $this->matches)) {
                                 $return[$title][] = [
-                                    'id' => (int) $this->matches[2],
+                                    'mal_id' => (int) $this->matches[2],
                                     'url' => BASE_URL . $this->matches[1] . '/' . $this->matches[2] . '/' . $this->matches[3],
                                     'title' => strip_tags($this->matches[4])
                                 ];